Fix code typo
Incredibly perplexing that this was not failing long ago with an error saying that .bbox does not exist, because it shouldn't. There is nowhere in the problem that sets `.bbox` on a WeatherModel object except in this one line in this function. That one line is the only possibly place it could be set, and yet I observed weather model objects _indeed_ having a .bbox property, and it _wasn't_ (-180, -90, 180, 90).
No idea what to make of that. The program is haunted. But this change makes this function use the property I believe it intended to use.