Hack: always project GeocodedFile to EPSG:4326
Rest of the code is not properly set up to handle GeocodedFile AOIs that are not in EPSG:4326, which causes problems with EPSG:32617 interferograms from ASF On Demand. I personally lack the time to fix this properly, but I did end up doing this much so that RAiDER would work with my code for the gemlab ASF On Demand notebook jlmaurer/gemlab#11.

In the case when the user inputs a geocoded file with projection other than EPSG:4326, this change fixes a crash, at the cost of probably more projection conversion artifacts than necessary from a more in-depth solution. In all other cases (other projections, other input AOI types), nothing is changed/affected.